Realtime-temperature-monitor-setup
The real-time temperature monitor setup contains will allow a user to see the real-time plotted temperature on a graph and the average value of the temperature. To visualize the graph we need some external hardware that connects to a specific comport that can be selected in the software.  The hardware required is Arduino & DHT11 sensor. Arduino will be connected to a COM port then the app will serially acquire the data from Arduino and plot it on the graph. You need a specific piece of code for the Arduino whose HEX file is attached you just need to upload that hex file on your Arduino board.